Config to db (#93)
* add cache

* add configs

* add configs

* add configs

* change config

* change config

* refactor: clean up method signatures and improve code formatting across multiple handlers

* delete message

---------

+++ DecembristChatBotSharp/Service/OpenBoxService.cs
@@ -1,4 +1,5 @@
 ﻿using DecembristChatBotSharp.Entity;
+using DecembristChatBotSharp.Entity.Configs;
 using DecembristChatBotSharp.Mongo;
 using Lamar;
 using Serilog;
@@ -9,25 +10,32 @@ namespace DecembristChatBotSharp.Service;
 public class OpenBoxService(
     MemberItemRepository memberItemRepository,
     MemberItemService memberItemService,
-    AppConfig appConfig,
     MongoDatabase db,
     HistoryLogRepository historyLogRepository,
     Random random,
     CancellationTokenSource cancelToken,
-    UniqueItemService uniqueItemService)
+    UniqueItemService uniqueItemService,
+    ChatConfigService chatConfigService)
 {
     public async Task<OpenBoxResultData> OpenBox(long chatId, long telegramId)
     {
+        var maybeConfig = await chatConfigService.GetConfig(chatId, config => config.ItemConfig);
+        if (!maybeConfig.TryGetSome(out var itemConfig))
+        {
+            return chatConfigService.LogNonExistConfig(
+                new OpenBoxResultData(MemberItemType.TelegramMeme, 0, OpenBoxResult.Failed), nameof(ItemConfig));
+        }
+
         using var session = await db.OpenSession();
         session.StartTransaction();
 
         var hasBox = await memberItemRepository.RemoveMemberItem(chatId, telegramId, MemberItemType.Box, session);
         if (!hasBox) return await AbortWithResult(session, OpenBoxResult.NoItems);
 
-        var (itemType, quantity) = GetRandomItemWithQuantity();
+        var (itemType, quantity) = GetRandomItemWithQuantity(itemConfig);
         return itemType switch
         {
-            MemberItemType.Stone => await HandleUniqueItem(chatId, telegramId, itemType, session),
+            MemberItemType.Stone => await HandleUniqueItem(chatId, telegramId, itemType, session, itemConfig),
             MemberItemType.Box =>
                 await HandleItemType(chatId, telegramId, itemType, quantity, OpenBoxResult.SuccessX2, session),
             MemberItemType.Amulet when await memberItemService.HandleAmuletItem((telegramId, chatId), session) =>
@@ -37,12 +45,12 @@ public class OpenBoxService(
     }
 
     private async Task<OpenBoxResultData> HandleUniqueItem(
-        long chatId, long telegramId, MemberItemType itemType, IMongoSession session)
+        long chatId, long telegramId, MemberItemType itemType, IMongoSession session, ItemConfig itemConfig)
     {
         var isHasUniqueItem = await memberItemRepository.IsUserHasItem(chatId, telegramId, itemType, session);
         if (isHasUniqueItem)
         {
-            var compensation = appConfig.ItemConfig.CompensationItem;
+            var compensation = itemConfig.CompensationItem;
             Log.Information("User has unique {0}, compensating item: {1} has been issued", itemType, compensation);
             return await LogInHistoryAndCommit(chatId, telegramId, OpenBoxResult.Success, compensation, session, 1);
         }
@@ -91,9 +99,9 @@ public class OpenBoxService(
         return new OpenBoxResultData(None, 0, result);
     }
 
-    private (MemberItemType, int) GetRandomItemWithQuantity()
+    private (MemberItemType, int) GetRandomItemWithQuantity(ItemConfig itemConfig)
     {
-        var itemChances = appConfig.ItemConfig.ItemChance;
+        var itemChances = itemConfig.ItemChance;
         var total = itemChances.Values.Sum(x => x.Chance);
         var roll = random.NextDouble() * total;
